added a check if the plugin already exists

+++ jenkins-master/plugins.sh
@@ -15,14 +15,17 @@ mkdir -p $REF
 
 while read spec || [ -n "$spec" ]; do
     plugin=(${spec//:/ });
-    [[ ${plugin[0]} =~ ^# ]] && continue
-    [[ ${plugin[0]} =~ ^\s*$ ]] && continue
-    [[ -z ${plugin[1]} ]] && plugin[1]="latest"
-    echo "Downloading ${plugin[0]}:${plugin[1]}"
+    if [ ! -f /var/jenkins_home/plugins/${plugin[0]}.jpi ]; then
+      [[ ${plugin[0]} =~ ^# ]] && continue
+      [[ ${plugin[0]} =~ ^\s*$ ]] && continue
+      [[ -z ${plugin[1]} ]] && plugin[1]="latest"
+      echo "Downloading ${plugin[0]}:${plugin[1]}"
 
-    if [ -z "$JENKINS_UC_DOWNLOAD" ]; then
-      JENKINS_UC_DOWNLOAD=$JENKINS_UC/download
+      if [ -z "$JENKINS_UC_DOWNLOAD" ]; then
+        JENKINS_UC_DOWNLOAD=$JENKINS_UC/download
+      fi
+      curl -sSL -f ${JENKINS_UC_DOWNLOAD}/plugins/${plugin[0]}/${plugin[1]}/${plugin[0]}.hpi -o $REF/${plugin[0]}.jpi
+      unzip -qqt $REF/${plugin[0]}.jpi
     fi
-    curl -sSL -f ${JENKINS_UC_DOWNLOAD}/plugins/${plugin[0]}/${plugin[1]}/${plugin[0]}.hpi -o $REF/${plugin[0]}.jpi
-    unzip -qqt $REF/${plugin[0]}.jpi
 done  < $1
+echo "All plugins up to date."
